rename tests folder to the more accurate usecases
